3/23
I am continuing the tutorial. It is taking longer than I expected, but by the end of the period tomorrow, I should be done with the tutorial and ready to move onto more advanced concepts in my project. The TypeScript to JavaScript conversion is confusing, but I have managed to fare well for the most part.